Refactor LoginListViewController to MVVM (again) (#6871)
* Refactor LoginListViewController to MVVM
* New PR made because previous changes were not merged correctly
