+++ Client/Frontend/Login Management/LoginListViewController.swift
@@ -8,14 +8,6 @@ import Storage
 import Shared
 import SwiftKeychainWrapper
 
-private struct LoginListUX {
-    static let RowHeight: CGFloat = 58
-    static let SearchHeight: CGFloat = 58
-    static let selectionButtonFont = UIFont.systemFont(ofSize: 16)
-    static let NoResultsFont = UIFont.systemFont(ofSize: 16)
-    static let NoResultsTextColor = UIColor.Photon.Grey40
-}
-
 private extension UITableView {
     var allLoginIndexPaths: [IndexPath] {
         return ((LoginsSettingsSection + 1)..<self.numberOfSections).flatMap { sectionNum in
@@ -26,25 +18,20 @@ private extension UITableView {
     }
 }
 
-private let CellReuseIdentifier = "cell-reuse-id"
-private let SectionHeaderId = "section-header-id"
-private let LoginsSettingsSection = 0
+let CellReuseIdentifier = "cell-reuse-id"
+let SectionHeaderId = "section-header-id"
+let LoginsSettingsSection = 0
 
 class LoginListViewController: SensitiveViewController {
 
-    fileprivate lazy var loginSelectionController: ListSelectionController = {
-        return ListSelectionController(tableView: self.tableView)
+    private let viewModel: LoginListViewModel
+
+    fileprivate lazy var loginSelectionController: LoginListSelectionHelper = {
+        return LoginListSelectionHelper(tableView: self.tableView)
     }()
 
-    fileprivate lazy var loginDataSource: LoginDataSource = {
-        let dataSource = LoginDataSource(profile: profile, searchController: searchController)
-        dataSource.dataObserver = self
-        return dataSource
-    }()
-
-    fileprivate let profile: Profile
+    fileprivate var loginDataSource: LoginDataSource
     fileprivate let searchController = UISearchController(searchResultsController: nil)
-    fileprivate var activeLoginQuery: Deferred<Maybe<[LoginRecord]>>?
     fileprivate let loadingView = SettingsLoadingView()
     fileprivate var deleteAlert: UIAlertController?
     fileprivate var selectionButtonHeightConstraint: Constraint?
@@ -61,7 +48,7 @@ class LoginListViewController: SensitiveViewController {
 
     fileprivate lazy var selectionButton: UIButton = {
         let button = UIButton()
-        button.titleLabel?.font = LoginListUX.selectionButtonFont
+        button.titleLabel?.font = LoginListViewModel.LoginListUX.selectionButtonFont
         button.addTarget(self, action: #selector(tappedSelectionButton), for: .touchUpInside)
         return button
     }()
@@ -109,7 +96,8 @@ class LoginListViewController: SensitiveViewController {
     }
 
     private init(profile: Profile) {
-        self.profile = profile
+        self.viewModel = LoginListViewModel(profile: profile, searchController: searchController)
+        self.loginDataSource = LoginDataSource(viewModel: self.viewModel)
         super.init(nibName: nil, bundle: nil)
     }
 
@@ -171,6 +159,7 @@ class LoginListViewController: SensitiveViewController {
         applyTheme()
 
         KeyboardHelper.defaultHelper.addDelegate(self)
+        viewModel.delegate = self
     }
 
     override func viewWillAppear(_ animated: Bool) {
@@ -238,21 +227,11 @@ class LoginListViewController: SensitiveViewController {
     }
 
     fileprivate func toggleSelectionTitle() {
-        if loginSelectionController.selectedCount == loginDataSource.count {
+        if loginSelectionController.selectedCount == viewModel.count {
             selectionButton.setTitle(deselectAllTitle, for: [])
         } else {
             selectionButton.setTitle(selectAllTitle, for: [])
         }
-    }
-
-    // Wrap the SQLiteLogins method to allow us to cancel it from our end.
-    fileprivate func queryLogins(_ query: String) -> Deferred<Maybe<[LoginRecord]>> {
-        let deferred = Deferred<Maybe<[LoginRecord]>>()
-        profile.logins.searchLoginsWithQuery(query) >>== { logins in
-            deferred.fillIfUnfilled(Maybe(success: logins.asArray()))
-            succeed()
-        }
-        return deferred
     }
 }
 
@@ -263,15 +242,13 @@ extension LoginListViewController: UISearchResultsUpdating {
     }
 }
 
-fileprivate var isDuringSearchControllerDismiss = false
-
 extension LoginListViewController: UISearchControllerDelegate {
     func willDismissSearchController(_ searchController: UISearchController) {
-        isDuringSearchControllerDismiss = true
+        viewModel.setIsDuringSearchControllerDismiss(to: true)
     }
 
     func didDismissSearchController(_ searchController: UISearchController) {
-        isDuringSearchControllerDismiss = false
+        viewModel.setIsDuringSearchControllerDismiss(to: false)
     }
 }
 
@@ -289,11 +266,7 @@ private extension LoginListViewController {
 
     func loadLogins(_ query: String? = nil) {
         loadingView.isHidden = false
-
-        // Fill in an in-flight query and re-query
-        activeLoginQuery?.fillIfUnfilled(Maybe(success: []))
-        activeLoginQuery = queryLogins(query ?? "")
-        activeLoginQuery! >>== loginDataSource.setLogins
+        viewModel.loadLogins(query, loginDataSource: self.loginDataSource)
     }
 
     @objc func beginEditing() {
@@ -320,14 +293,14 @@ private extension LoginListViewController {
     }
 
     @objc func tappedDelete() {
-        profile.logins.hasSyncedLogins().uponQueue(.main) { yes in
+        viewModel.profile.logins.hasSyncedLogins().uponQueue(.main) { yes in
             self.deleteAlert = UIAlertController.deleteLoginAlertWithDeleteCallback({ [unowned self] _ in
                 // Delete here
                 let guidsToDelete = self.loginSelectionController.selectedIndexPaths.compactMap { indexPath in
-                    self.loginDataSource.loginAtIndexPath(indexPath)?.id
+                    self.viewModel.loginAtIndexPath(indexPath)?.id
                 }
 
-                self.profile.logins.delete(ids: guidsToDelete).uponQueue(.main) { _ in
+                self.viewModel.profile.logins.delete(ids: guidsToDelete).uponQueue(.main) { _ in
                     self.cancelSelection()
                     self.loadLogins()
                 }
@@ -339,7 +312,7 @@ private extension LoginListViewController {
 
     @objc func tappedSelectionButton() {
         // If we haven't selected everything yet, select all
-        if loginSelectionController.selectedCount < loginDataSource.count {
+        if loginSelectionController.selectedCount < viewModel.count {
             // Find all unselected indexPaths
             let unselectedPaths = tableView.allLoginIndexPaths.filter { indexPath in
                 return !loginSelectionController.indexPathIsSelected(indexPath)
@@ -360,23 +333,6 @@ private extension LoginListViewController {
 
         toggleSelectionTitle()
         toggleDeleteBarButton()
-    }
-}
-
-// MARK: - LoginDataSourceObserver
-extension LoginListViewController: LoginDataSourceObserver {
-    func loginSectionsDidUpdate() {
-        loadingView.isHidden = true
-        tableView.reloadData()
-        activeLoginQuery = nil
-        navigationItem.rightBarButtonItem?.isEnabled = loginDataSource.count > 0
-        restoreSelectedRows()
-    }
-
-    func restoreSelectedRows() {
-        for path in self.loginSelectionController.selectedIndexPaths {
-            tableView.selectRow(at: path, animated: false, scrollPosition: .none)
-        }
     }
 }
 
@@ -407,7 +363,7 @@ extension LoginListViewController: UITableViewDelegate {
         if indexPath.section == LoginsSettingsSection, searchController.isActive || tableView.isEditing {
             return 0
         }
-        return indexPath.section == LoginsSettingsSection ? 44 : LoginListUX.RowHeight
+        return indexPath.section == LoginsSettingsSection ? 44 : LoginListViewModel.LoginListUX.RowHeight
     }
 
     func tableView(_ tableView: UITableView, editingStyleForRowAt indexPath: IndexPath) -> UITableViewCell.EditingStyle {
@@ -419,9 +375,9 @@ extension LoginListViewController: UITableViewDelegate {
             loginSelectionController.selectIndexPath(indexPath)
             toggleSelectionTitle()
             toggleDeleteBarButton()
-        } else if let login = loginDataSource.loginAtIndexPath(indexPath) {
+        } else if let login = viewModel.loginAtIndexPath(indexPath) {
             tableView.deselectRow(at: indexPath, animated: true)
-            let detailViewController = LoginDetailViewController(profile: profile, login: login)
+            let detailViewController = LoginDetailViewController(profile: viewModel.profile, login: login)
             detailViewController.settingsDelegate = settingsDelegate
             navigationController?.pushViewController(detailViewController, animated: true)
         }
@@ -474,280 +430,19 @@ extension LoginListViewController: SearchInputViewDelegate {
     }
 }
 
-/// Controller that keeps track of selected indexes
-fileprivate class ListSelectionController: NSObject {
-    private unowned let tableView: UITableView
-    private(set) var selectedIndexPaths = [IndexPath]()
+// MARK: - LoginViewModelDelegate
+extension LoginListViewController: LoginViewModelDelegate {
 
-    var selectedCount: Int {
-        return selectedIndexPaths.count
+    func loginSectionsDidUpdate() {
+        loadingView.isHidden = true
+        tableView.reloadData()
+        navigationItem.rightBarButtonItem?.isEnabled = viewModel.count > 0
+        restoreSelectedRows()
     }
 
-    init(tableView: UITableView) {
-        self.tableView = tableView
-        super.init()
-    }
-
-    func selectIndexPath(_ indexPath: IndexPath) {
-        selectedIndexPaths.append(indexPath)
-    }
-
-    func indexPathIsSelected(_ indexPath: IndexPath) -> Bool {
-        return selectedIndexPaths.contains(indexPath) { path1, path2 in
-            return path1.row == path2.row && path1.section == path2.section
+    func restoreSelectedRows() {
+        for path in self.loginSelectionController.selectedIndexPaths {
+            tableView.selectRow(at: path, animated: false, scrollPosition: .none)
         }
-    }
-
-    func deselectIndexPath(_ indexPath: IndexPath) {
-        guard let foundSelectedPath = (selectedIndexPaths.filter { $0.row == indexPath.row && $0.section == indexPath.section }).first,
-              let indexToRemove = selectedIndexPaths.firstIndex(of: foundSelectedPath) else {
-            return
-        }
-
-        selectedIndexPaths.remove(at: indexToRemove)
-    }
-
-    func deselectAll() {
-        selectedIndexPaths.removeAll()
-    }
-
-    func selectIndexPaths(_ indexPaths: [IndexPath]) {
-        selectedIndexPaths += indexPaths
-    }
-}
-
-protocol LoginDataSourceObserver: AnyObject {
-    func loginSectionsDidUpdate()
-}
-
-/// Data source for handling LoginData objects from a Cursor
-class LoginDataSource: NSObject, UITableViewDataSource {
-    var count = 0
-    weak var dataObserver: LoginDataSourceObserver?
-    weak var searchController: UISearchController?
-    fileprivate let emptyStateView = NoLoginsView()
-    fileprivate var titles = [Character]()
-
-    let boolSettings: (BoolSetting, BoolSetting)
-
-    init(profile: Profile, searchController: UISearchController) {
-        self.searchController = searchController
-        boolSettings = (
-            BoolSetting(prefs: profile.prefs, prefKey: PrefsKeys.LoginsSaveEnabled, defaultValue: true, attributedTitleText: NSAttributedString(string: Strings.SettingToSaveLogins)),
-            BoolSetting(prefs: profile.prefs, prefKey: PrefsKeys.LoginsShowShortcutMenuItem, defaultValue: true, attributedTitleText: NSAttributedString(string: Strings.SettingToShowLoginsInAppMenu)))
-        super.init()
-    }
-
-    fileprivate var loginRecordSections = [Character: [LoginRecord]]() {
-        didSet {
-            assert(Thread.isMainThread)
-            self.dataObserver?.loginSectionsDidUpdate()
-        }
-    }
-
-    fileprivate func loginsForSection(_ section: Int) -> [LoginRecord]? {
-        guard section > 0 else {
-            assertionFailure()
-            return nil
-        }
-        let titleForSectionIndex = titles[section - 1]
-        return loginRecordSections[titleForSectionIndex]
-    }
-
-    func loginAtIndexPath(_ indexPath: IndexPath) -> LoginRecord? {
-        guard indexPath.section > 0 else {
-            assertionFailure()
-            return nil
-        }
-        let titleForSectionIndex = titles[indexPath.section - 1]
-        guard let section = loginRecordSections[titleForSectionIndex] else {
-            assertionFailure()
-            return nil
-        }
-
-        assert(indexPath.row <= section.count)
-
-        return section[indexPath.row]
-    }
-
-    @objc func numberOfSections(in tableView: UITableView) -> Int {
-        if  loginRecordSections.isEmpty {
-            tableView.backgroundView = emptyStateView
-            return 1
-        }
-
-        tableView.backgroundView = nil
-        // Add one section for the settings section.
-        return loginRecordSections.count + 1
-    }
-
-    @objc func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        if section == LoginsSettingsSection {
-            return 2
-        }
-        return loginsForSection(section)?.count ?? 0
-    }
-
-    @objc func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = ThemedTableViewCell(style: .subtitle, reuseIdentifier: CellReuseIdentifier)
-
-        if indexPath.section == LoginsSettingsSection {
-            let hideSettings = searchController?.isActive ?? false || tableView.isEditing
-            let setting = indexPath.row == 0 ? boolSettings.0 : boolSettings.1
-            setting.onConfigureCell(cell)
-            if hideSettings {
-                cell.isHidden = true
-            }
-
-            // Fade in the cell while dismissing the search or the cell showing suddenly looks janky
-            if isDuringSearchControllerDismiss {
-                cell.isHidden = false
-                cell.contentView.alpha = 0
-                cell.accessoryView?.alpha = 0
-                UIView.animate(withDuration: 0.6) {
-                    cell.contentView.alpha = 1
-                    cell.accessoryView?.alpha = 1
-                }
-            }
-        } else {
-            guard let login = loginAtIndexPath(indexPath) else { return cell }
-            cell.textLabel?.text = login.hostname
-            cell.detailTextColor = UIColor.theme.tableView.rowDetailText
-            cell.detailTextLabel?.text = login.username
-            cell.accessoryType = .disclosureIndicator
-        }
-        
-        // Need to override the default background multi-select color to support theming
-        cell.multipleSelectionBackgroundView = UIView()
-        cell.applyTheme()
-        return cell
-    }
-
-    func setLogins(_ logins: [LoginRecord]) {
-        // NB: Make sure we call the callback on the main thread so it can be synced up with a reloadData to
-        //     prevent race conditions between data/UI indexing.
-        return computeSectionsFromLogins(logins).uponQueue(.main) { result in
-            guard let (titles, sections) = result.successValue else {
-                self.count = 0
-                self.titles = []
-                self.loginRecordSections = [:]
-                return
-            }
-
-            self.count = logins.count
-            self.titles = titles
-            self.loginRecordSections = sections
-
-            // Disable the search controller if there are no logins saved
-            if !(self.searchController?.isActive ?? true) {
-                self.searchController?.searchBar.isUserInteractionEnabled = !logins.isEmpty
-                self.searchController?.searchBar.alpha = logins.isEmpty ? 0.5 : 1.0
-            }
-        }
-    }
-
-    fileprivate func computeSectionsFromLogins(_ logins: [LoginRecord]) -> Deferred<Maybe<([Character], [Character: [LoginRecord]])>> {
-        guard logins.count > 0 else {
-            return deferMaybe( ([Character](), [Character: [LoginRecord]]()) )
-        }
-
-        var domainLookup = [GUID: (baseDomain: String?, host: String?, hostname: String)]()
-        var sections = [Character: [LoginRecord]]()
-        var titleSet = Set<Character>()
-
-        // Small helper method for using the precomputed base domain to determine the title/section of the
-        // given login.
-        func titleForLogin(_ login: LoginRecord) -> Character {
-            // Fallback to hostname if we can't extract a base domain.
-            let titleString = domainLookup[login.id]?.baseDomain?.uppercased() ?? login.hostname
-            return titleString.first ?? Character("")
-        }
-
-        // Rules for sorting login URLS:
-        // 1. Compare base domains
-        // 2. If bases are equal, compare hosts
-        // 3. If login URL was invalid, revert to full hostname
-        func sortByDomain(_ loginA: LoginRecord, loginB: LoginRecord) -> Bool {
-            guard let domainsA = domainLookup[loginA.id],
-                  let domainsB = domainLookup[loginB.id] else {
-                return false
-            }
-
-            guard let baseDomainA = domainsA.baseDomain,
-                  let baseDomainB = domainsB.baseDomain,
-                  let hostA = domainsA.host,
-                let hostB = domainsB.host else {
-                return domainsA.hostname < domainsB.hostname
-            }
-
-            if baseDomainA == baseDomainB {
-                return hostA < hostB
-            } else {
-                return baseDomainA < baseDomainB
-            }
-        }
-
-        return deferDispatchAsync(DispatchQueue.global(qos: DispatchQoS.userInteractive.qosClass)) {
-            // Precompute the baseDomain, host, and hostname values for sorting later on. At the moment
-            // baseDomain() is a costly call because of the ETLD lookup tables.
-            logins.forEach { login in
-                domainLookup[login.id] = (
-                    login.hostname.asURL?.baseDomain,
-                    login.hostname.asURL?.host,
-                    login.hostname
-                )
-            }
-
-            // 1. Temporarily insert titles into a Set to get duplicate removal for 'free'.
-            logins.forEach { titleSet.insert(titleForLogin($0)) }
-
-            // 2. Setup an empty list for each title found.
-            titleSet.forEach { sections[$0] = [LoginRecord]() }
-
-            // 3. Go through our logins and put them in the right section.
-            logins.forEach { sections[titleForLogin($0)]?.append($0) }
-
-            // 4. Go through each section and sort.
-            sections.forEach { sections[$0] = $1.sorted(by: sortByDomain) }
-
-            return deferMaybe( (Array(titleSet).sorted(), sections) )
-        }
-    }
-}
-
-// Empty state view when there is no logins to display.
-fileprivate class NoLoginsView: UIView {
-
-    // We use the search bar height to maintain visual balance with the whitespace on this screen. The
-    // title label is centered visually using the empty view + search bar height as the size to center with.
-    var searchBarHeight: CGFloat = 0 {
-        didSet {
-            setNeedsUpdateConstraints()
-        }
-    }
-
-    lazy var titleLabel: UILabel = {
-        let label = UILabel()
-        label.font = LoginListUX.NoResultsFont
-        label.textColor = LoginListUX.NoResultsTextColor
-        label.text = NSLocalizedString("No logins found", tableName: "LoginManager", comment: "Label displayed when no logins are found after searching.")
-        return label
-    }()
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        addSubview(titleLabel)
-    }
-
-    fileprivate override func updateConstraints() {
-        super.updateConstraints()
-        titleLabel.snp.remakeConstraints { make in
-            make.centerX.equalTo(self)
-            make.centerY.equalTo(self).offset(-(searchBarHeight / 2))
-        }
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
     }
 }
